Updated pyChromVar functionality
Added:
- get_bg_peaks (deterministic KNN for chromVAR's real binned gaussian
  weighed sampling)
- compute_deviations (now emits z-scores and deviations + QC) also fixed
  background to reoder cound cols
- match_motif (implemented TODO sub there)

Deps:
- added pytest and pytest-xdist, dropped pynndescent, removed deprecated
  anndata APIs

+++ tests/test_preprocessing.py
@@ -4,7 +4,7 @@ from anndata import AnnData
 from pychromvar.preprocessing import add_gc_bias
 
 def test_gc_bias():
-    data = AnnData(np.array([[1, 2,], [3, 4]]), dtype=np.float32)
+    data = AnnData(np.array([[1, 2,], [3, 4]]))
     data.uns["peak_seq"] = ["AAAAAAAA", "CCCCGGGG"]
     add_gc_bias(data=data)
